refactor: safe parallelism, tech debt deferral, graph staleness fallback
- Relax mandatory task creation requirement; permit read-only exploration before TaskCreate
- Adjust parallelism guidance: sequential by default for writes, parallel only for isolated domains
- Soften bug-fix mandate: track bugs as low-priority tasks, defer at session end to prevent feature starvation
- Update Impact Radius Protocol: Category A bugs tracked for end-of-session fixing, not immediate fixes
- Clarify MCP tool usage: prioritize but fallback gracefully if services fail
- Fix DuckDB schema: use DOUBLE instead of REAL for numeric columns (mtime, size)
- Update test assertion: Stop hook graph update now at index 2 (after error logging hooks)
- Align documentation across CLAUDE.md, README, agents, and skills with new workflow philosophy

+++ src/graph/db/schema.ts
@@ -43,8 +43,8 @@ export async function ensureSchema(db: DuckDBConnection): Promise<void> {
     CREATE TABLE IF NOT EXISTS scan_cache (
       path TEXT PRIMARY KEY,
       hash TEXT NOT NULL,
-      mtime REAL NOT NULL,
-      size REAL,
+      mtime DOUBLE NOT NULL,
+      size DOUBLE,
       language TEXT,
       scannerVersion TEXT,
       nodes TEXT NOT NULL,
